requirements.txt
google-genai
python-dotenv
rich
tiktoken
